Safari-611.1.10/LayoutTests/fast/events/wheel/wheelevent-direction-inverted-from-device-expected.txt

This test checks for the existence of the webkitDirectionInvertedFromDevice property in the WheelEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Scroll over me
sending scroll event to green target
PASS result != undefined is true
PASS successfullyParsed is true

TEST COMPLETE
